requirements.txt
astroid==2.2.3
certifi==2018.11.29
chardet==3.0.4
colorama==0.4.1
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.7
django-rest-swagger==2.2.0
djangorestframework==3.9.2
gunicorn==19.9.0
idna==2.8
isort==4.3.10
itypes==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.1
mccabe==0.6.1
openapi-codec==1.3.2
Pillow==5.4.1
pylint==2.3.1
pytz==2018.9
requests==2.21.0
simplejson==3.16.0
six==1.12.0
typed-ast==1.3.1
uritemplate==3.0.0
urllib3==1.24.1
whitenoise==4.1.2
wrapt==1.11.1
